layers/knowledge/api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.23.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==45.0.5
dataclasses-json==0.6.7
defusedxml==0.7.1
dill==0.4.0
docling==2.40.0
docling-core==2.40.0
docling-ibm-models==3.7.0
docling-parse==4.1.0
easyocr==1.7.2
et-xmlfile==2.0.0
fastapi==0.104.1
ffmpeg-downloader==0.4.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
isodate==0.7.2
jinja2==3.1.6
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-ollama==0.3.5
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.4.8
latex2mathml==3.78.0
lazy-loader==0.4
lxml==5.4.0
magika==0.6.2
mammoth==1.9.1
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.2
marko==2.1.4
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
minio==7.2.15
mpire==2.10.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
multiprocess==0.70.18
mypy-extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
olefile==0.47
ollama==0.5.1
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.0
packaging==25.0
pandas==2.3.0
pdfminer-six==20250506
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
portalocker==2.10.1
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
pyjwt==2.10.1
pylatexenc==2.10
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
semchunk==2.2.2
sentence-transformers==5.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
speechrecognition==3.14.3
sqlalchemy==2.0.41
starlette==0.27.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
zstandard==0.23.0
